pyproject.toml
[project]
name = "weebiee-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.72.1",
    "grpcio-tools>=1.72.1",
    "pandas>=2.3.0",
    "pyspark-client>=4.0.0",
    "torch>=2.7.1",
    "transformers>=4.52.4",
]
